Extracts insight as hook
